cfl: add CFL-062..066 UBSAN patches + fix analyzer overflow
CFL-062: icGetSig implicit char conversion (4 sites, CWE-681)
CFL-063: bounds check unsigned overflow (4 sites, CWE-190)
CFL-064: segmented curve subtraction underflow (2 sites, CWE-191)
CFL-065: IccTagLut nEnd subtraction underflow (12 sites, CWE-191)
CFL-066: IccUtilXml bitmask signed conversion (1 site, CWE-681)
Fix: IccHeuristicsCodeQLPatterns.cpp:925 uint32 overflow guard

+++ iccanalyzer-lite/IccHeuristicsCodeQLPatterns.cpp
@@ -922,7 +922,7 @@ int RunHeuristic_H173_SigConversionShiftOverflow(RawProfileContext &ctx)
 
   // Tag type signatures (first 4 bytes of each tag's data)
   for (const auto &t : ctx.tags) {
-    if (t.size < 4 || t.offset + 4 > ctx.fileSize()) continue;
+    if (t.size < 4 || (uint64_t)t.offset + 4 > ctx.fileSize()) continue;
     uint8_t buf[4];
     if (!ctx.ReadAt(t.offset, buf, 4)) continue;
     uint32_t typeSig = ReadU32BE(buf);
